refactor(bspline): split bspline_space.py into 1D, ND, and factory modules
