+++ src/pantr/bspline_space_1D.py
@@ -1,16 +1,14 @@
-"""BsplineSpace class and  utilities."""
+"""BsplineSpace1D class for 1D B-spline spaces."""
 
 from __future__ import annotations
 
 import functools
-from collections.abc import Iterable
 from typing import TYPE_CHECKING
 
 import numpy as np
 from numpy import typing as npt
 
 from ._bspline_basis_core import _tabulate_Bspline_basis_1D_impl
-from ._bspline_basis_multidim import _tabulate_Bspline_basis_impl
 from ._bspline_extraction import (
     _tabulate_Bspline_Bezier_1D_extraction_impl,
     _tabulate_Bspline_cardinal_1D_extraction_impl,
@@ -19,224 +17,12 @@ from ._bspline_extraction import (
 from ._bspline_knots import (
     _get_Bspline_cardinal_intervals_1D_impl,
     _get_Bspline_num_basis_1D_impl,
-    _get_knots_ends_and_dtype,
     _get_unique_knots_and_multiplicity_impl,
-    _validate_knot_input,
 )
 from .tolerance import get_strict_tolerance
 
 if TYPE_CHECKING:
-    from .quad import PointsLattice
-
-
-def create_uniform_open_knot_vector(
-    num_intervals: int,
-    degree: int,
-    continuity: int | None = None,
-    domain: tuple[np.float32 | np.float64 | float, np.float32 | np.float64 | float] | None = None,
-    dtype: npt.DTypeLike | None = None,
-) -> npt.NDArray[np.float32 | np.float64]:
-    """Create a uniform open knot vector.
-
-    An open knot vector has the first and last knots repeated (degree+1) times,
-    ensuring the B-spline interpolates the first and last control points.
-
-    Args:
-        num_intervals (int): Number of intervals in the domain. Must be non-negative.
-        degree (int): B-spline degree. Must be non-negative.
-        continuity (Optional[int]): Continuity level at interior knots.
-            Must be between -1 and degree-1. Defaults to degree-1 (maximum continuity).
-        domain (Optional[tuple[np.float32 | np.float64 | float, np.float32 | np.float64 | float]]):
-            Domain boundaries as (start, end). Defaults to (0.0, 1.0) if not provided.
-        dtype (Optional[np.dtype]): Data type for the knot vector.
-            If None, inferred from start/end or defaults to float64.
-
-    Returns:
-        npt.NDArray[np.floating]: Open knot vector with uniform spacing.
-
-    Raises:
-        ValueError: If any parameter is invalid.
-
-    Example:
-        >>> create_uniform_open_knot_vector(2, 2, domain=(0.0, 1.0))
-        array([0., 0., 0., 0.5, 1., 1., 1.])
-    """
-    start_value: np.float32 | np.float64 | None
-    end_value: np.float32 | np.float64 | None
-    if domain is None:
-        start_value = None
-        end_value = None
-    else:
-        start_raw, end_raw = domain
-        start_value = start_raw if isinstance(start_raw, np.floating) else np.float64(start_raw)
-        end_value = end_raw if isinstance(end_raw, np.floating) else np.float64(end_raw)
-
-    start, end, dtype = _get_knots_ends_and_dtype(start_value, end_value, dtype)
-
-    continuity = degree - 1 if continuity is None else continuity
-
-    _validate_knot_input(
-        num_intervals,
-        degree,
-        continuity,
-        (start, end),
-        dtype,
-    )
-
-    unique_knots = np.linspace(start, end, num_intervals + 1, dtype=dtype)
-    knots = np.array([start] * (degree + 1), dtype)
-
-    interior_multiplicity = degree - continuity
-    for knot in unique_knots[1:-1]:
-        knots = np.append(knots, [knot] * interior_multiplicity)
-
-    knots = np.append(knots, [end] * (degree + 1))
-
-    return knots
-
-
-def create_uniform_periodic_knot_vector(
-    num_intervals: int,
-    degree: int,
-    continuity: int | None = None,
-    domain: tuple[np.float32 | np.float64 | float, np.float32 | np.float64 | float] | None = None,
-    dtype: npt.DTypeLike | None = np.float64,
-) -> npt.NDArray[np.float32 | np.float64]:
-    """Create a uniform periodic knot vector.
-
-    A periodic knot vector extends beyond the domain boundaries to ensure
-    periodicity of the B-spline basis functions.
-
-    Args:
-        num_intervals (int): Number of intervals in the domain. Must be non-negative.
-        degree (int): B-spline degree. Must be non-negative.
-        continuity (Optional[int]): Continuity level at interior knots.
-            Must be between -1 and degree-1. Defaults to degree-1 (maximum continuity).
-        domain (Optional[tuple[np.float32 | np.float64 | float, np.float32 | np.float64 | float]]):
-            Domain boundaries as (start, end). Defaults to (0.0, 1.0) if not provided.
-        dtype (Optional[np.dtype]): Data type for the knot vector.
-            If None, inferred from start/end or defaults to float64.
-
-    Returns:
-        npt.NDArray[np.floating]: Periodic knot vector with uniform spacing.
-
-    Raises:
-        ValueError: If any parameter is invalid.
-
-    Example:
-        >>> create_uniform_periodic_knot_vector(2, 2, domain=(0.0, 1.0))
-        array([-1. , -0.5,  0. ,  0.5,  1. ,  1.5,  2. ])
-    """
-    start_value: np.float32 | np.float64 | None
-    end_value: np.float32 | np.float64 | None
-    if domain is None:
-        start_value = None
-        end_value = None
-    else:
-        start_raw, end_raw = domain
-        start_value = start_raw if isinstance(start_raw, np.floating) else np.float64(start_raw)
-        end_value = end_raw if isinstance(end_raw, np.floating) else np.float64(end_raw)
-
-    start, end, dtype = _get_knots_ends_and_dtype(start_value, end_value, dtype)
-    continuity = degree - 1 if continuity is None else continuity
-
-    _validate_knot_input(
-        num_intervals,
-        degree,
-        continuity,
-        (start, end),
-        dtype,
-    )
-
-    # Create uniform spacing for unique interior knots
-    unique_knots = np.linspace(start, end, num_intervals + 1, dtype=dtype)
-
-    # Build knot vector with repetitions
-    knots = np.array([], dtype=dtype)
-
-    multiplicity = degree - continuity
-
-    # Starting periodic knots.
-    length = (end - start) / num_intervals
-    knots = np.linspace(
-        start - length * (degree - multiplicity + 1),
-        start,
-        degree + 2 - multiplicity,
-        dtype=dtype,
-    )[:-1]
-
-    # Interior knots with specified multiplicity
-    for knot in unique_knots:
-        knots = np.append(knots, [knot] * multiplicity)
-
-    # End periodic knots.
-    knots = np.append(
-        knots,
-        np.linspace(
-            end,
-            end + length * (degree - multiplicity + 1),
-            degree + 2 - multiplicity,
-            dtype=dtype,
-        )[1:],
-    )
-
-    return knots
-
-
-def create_cardinal_Bspline_knot_vector(
-    num_intervals: int,
-    degree: int,
-    dtype: npt.DTypeLike = np.float64,
-) -> npt.NDArray[np.float32 | np.float64]:
-    """Create a knot vector for cardinal B-spline basis functions.
-
-    Cardinal B-splines are B-splines defined on uniform knot vectors with
-    maximum continuity, where the basis functions in the central region
-    have the same shape and are translated versions of each other.
-
-    Args:
-        num_intervals (int): Number of intervals in the domain. Must be at least 1.
-        degree (int): B-spline degree. Must be non-negative.
-        dtype (npt.DTypeLike): Data type for the knot vector.
-            It must be either float32 or float64. Defaults to np.float64.
-
-    Returns:
-        npt.NDArray[np.float32 | np.float64]: Cardinal B-spline knot vector
-            with uniform spacing.
-
-    Raises:
-        ValueError: If num_intervals < 1, degree < 0, or dtype is not float32/float64.
-
-    Example:
-        >>> create_cardinal_Bspline_knot_vector(2, 2)
-        array([-2., -1.,  0.,  1.,  2.,  3., 4.])
-    """
-    if num_intervals < 1:
-        raise ValueError("num_intervals must be at least 1")
-
-    if degree < 0:
-        raise ValueError("degree must be non-negative")
-
-    dtype_obj = np.dtype(dtype)
-    if dtype_obj not in (np.dtype(np.float32), np.dtype(np.float64)):
-        raise ValueError("dtype must be float32 or float64")
-
-    start_value: np.float32 | np.float64
-    end_value: np.float32 | np.float64
-    if dtype_obj == np.dtype(np.float64):
-        start_value = np.float64(0)
-        end_value = np.float64(num_intervals)
-    else:
-        start_value = np.float32(0)
-        end_value = np.float32(num_intervals)
-
-    return create_uniform_periodic_knot_vector(
-        num_intervals,
-        degree,
-        continuity=degree - 1,
-        domain=(start_value, end_value),
-        dtype=dtype_obj,
-    )
+    pass
 
 
 @functools.cache
@@ -746,201 +532,4 @@ class BsplineSpace1D:
         """
         return _tabulate_Bspline_cardinal_1D_extraction_impl(
             self.knots, self.degree, self.tolerance, out=out
-        )
-
-
-class BsplineSpace:
-    """A class representing a multi-dimensional B-spline space.
-
-    This space is defined by a set of B-spline spaces, one for each dimension.
-
-    This class provides methods to analyze B-spline properties, validate input
-    parameters, compute various geometric characteristics of the spline,
-    and access various properties of the B-spline.
-
-    Attributes:
-        _spaces (Iterable[BsplineSpace1D]): List of B-spline spaces, one for each dimension.
-    """
-
-    _spaces: tuple[BsplineSpace1D, ...]
-
-    def __init__(
-        self,
-        spaces: Iterable[BsplineSpace1D],
-    ) -> None:
-        """Initialize a B-spline space object.
-
-        Args:
-            spaces (Iterable[BsplineSpace1D]): List of B-spline spaces, one for each dimension.
-
-        Raises:
-            ValueError: If the B-spline spaces have different data types.
-        """
-        self._spaces = tuple(spaces)
-
-        if not all(space.dtype == self.dtype for space in self._spaces):
-            raise ValueError("All B-spline spaces must have the same data type.")
-
-    @property
-    def dim(self) -> int:
-        """Get the dimension of the B-spline space.
-
-        Returns:
-            int: The dimension of the B-spline space.
-        """
-        return len(self._spaces)
-
-    @property
-    def spaces(self) -> tuple[BsplineSpace1D, ...]:
-        """Get the B-spline spaces.
-
-        Returns:
-            tuple[BsplineSpace1D, ...]: The B-spline spaces.
-        """
-        return self._spaces
-
-    @functools.cached_property
-    def degrees(self) -> tuple[int, ...]:
-        """Get the polynomial degree of the B-spline.
-
-        Returns:
-            tuple[int, ...]: The degree for each dimension.
-        """
-        return tuple(space.degree for space in self._spaces)
-
-    @functools.cached_property
-    def tolerance(self) -> float:
-        """Get the tolerance value used for numerical comparisons.
-
-        It is the maximum tolerance of the B-spline spaces.
-
-        Returns:
-            float: The tolerance value.
-        """
-        return max(space.tolerance for space in self._spaces)
-
-    @property
-    def dtype(self) -> npt.DTypeLike:
-        """Get the data type of the B-spline space.
-
-        Returns:
-            npt.DTypeLike: The numpy data type of the B-spline space.
-        """
-        return self._spaces[0].dtype
-
-    @functools.cached_property
-    def num_basis(self) -> tuple[int, ...]:
-        """Get the number of basis functions for each dimension.
-
-        Returns:
-            tuple[int, ...]: The number of basis functions for each dimension.
-        """
-        return tuple(space.num_basis for space in self._spaces)
-
-    @functools.cached_property
-    def num_total_basis(self) -> int:
-        """Get the total number of basis functions.
-
-        Returns:
-            int: The total number of basis functions.
-        """
-        return int(np.prod(self.num_basis))
-
-    @functools.cached_property
-    def num_intervals(self) -> tuple[int, ...]:
-        """Get the number of intervals for each dimension.
-
-        Returns:
-            tuple[int, ...]: The number of intervals for each dimension.
-        """
-        return tuple(space.num_intervals for space in self._spaces)
-
-    @functools.cached_property
-    def num_total_intervals(self) -> int:
-        """Get the total number of intervals.
-
-        Returns:
-            int: The total number of intervals.
-        """
-        return int(np.prod(self.num_intervals))
-
-    @functools.cached_property
-    def domain(self) -> npt.NDArray[np.float32 | np.float64]:
-        """Get the domain of the B-spline space.
-
-        Returns:
-            npt.NDArray[np.float32 | np.float64]: The domain of the B-spline space.
-            The shape is (dim, 2), where the last dimension contains the start
-            and end values of the domain.
-        """
-        domain = np.empty((self.dim, 2), dtype=self.dtype)
-        for i, space in enumerate(self._spaces):
-            domain[i, :] = space.domain
-        return domain
-
-    def has_Bezier_like_knots(self) -> bool:
-        """Check if the knot vector represents a Bézier-like configuration.
-
-        A Bézier-like configuration has open ends and only one non-zero span
-        for each dimension.
-
-        Returns:
-            bool: True if knots have open ends and only one span.
-
-        Example:
-            >>> bspline_1D = BsplineSpace1D([1, 1, 1, 3, 3, 3], 2)
-            >>> bspline_2D = BsplineSpace([bspline_1D, bspline_1D])
-            >>> bspline_2D.has_Bezier_like_knots()
-            True
-        """
-        return all(space.has_Bezier_like_knots() for space in self._spaces)
-
-    def tabulate_basis(
-        self,
-        pts: npt.NDArray[np.float32 | np.float64] | PointsLattice,
-        out_basis: npt.NDArray[np.float32 | np.float64] | None = None,
-        out_first_basis: npt.NDArray[np.int_] | None = None,
-    ) -> tuple[npt.NDArray[np.float32 | np.float64], npt.NDArray[np.int_]]:
-        """Tabulate the B-spline basis functions at the given points.
-
-        Args:
-            pts (npt.NDArray[np.float32 | np.float64] | PointsLattice): The points
-               at which to tabulate the basis functions.
-               It can be a 2D array with shape (num_pts, dim) or a PointsLattice object.
-            out_basis (npt.NDArray[np.float32 | np.float64] | None): Optional output array where the
-                basis values will be stored. If None, a new array is allocated. Must have the
-                correct shape and dtype if provided. This follows NumPy's style for output arrays.
-                Defaults to None.
-            out_first_basis (npt.NDArray[np.int_] | None): Optional output array where the
-                first basis indices will be stored. If None, a new array is allocated. Must have
-                the correct shape and dtype np.int_ if provided. This follows NumPy's style for
-                output arrays. Defaults to None.
-
-        Returns:
-            tuple[npt.NDArray[np.float32 | np.float64], npt.NDArray[np.int_]]: The basis
-            function values and the first basis function indices.
-
-            In the case pts is a 2D array, the shape of the basis function values array
-            is (num_pts, order[0], order[1], ..., order[d-1]), where d is the dimension
-            of the B-spline space and num_pts is the number of points.
-            In the case pts is a PointsLattice object, the shape of the
-            basis function values array is
-            (num_pts_0, num_pts_1, ..., num_pts_d, order[0], order[1], ..., order[d-1]),
-            where num_pts_i is the number of points in the i-th dimension.
-
-            The shape of the first basis function indices array is (num_pts, dim),
-            if pts is a 2D array, or (num_pts_0, num_pts_1, ..., num_pts_d, dim),
-            if pts is a PointsLattice object.
-
-            If `out_basis` or `out_first_basis` was provided, the corresponding element of the tuple
-            is the same array.
-
-        Raises:
-            ValueError: If pts is not a 2D array or a PointsLattice object.
-            ValueError: If the pts dimension does not match the dimension of the B-spline space.
-            ValueError: If one or more points are outside the domain of the B-spline space, or if
-                `out_basis` or `out_first_basis` is provided and has incorrect shape or dtype.
-        """
-        return _tabulate_Bspline_basis_impl(
-            self, pts, out_basis=out_basis, out_first_basis=out_first_basis
         )
